#2 Add logout action and move player_new to SecurityController
